# PYTHON/api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "title": "Flask API",
      "description": "API for user authentication and account management"
    },
    "paths": {
      "/login": {
        "post": {
          "summary": "Login with username and password",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "headers": {
                "Set-Cookie": {
                  "description": "Authentication cookie",
                  "schema": {
                    "type": "string"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/email": {
        "get": {
          "summary": "Get email of current session",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/password/reset/request": {
        "post": {
          "summary": "Request password reset",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/account/login/code": {
        "post": {
          "summary": "Login with email and code",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCodeRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/password/reset": {
        "post": {
          "summary": "Reset password",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordReset"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/password/update": {
        "post": {
          "summary": "Update password",
          "security": [
            {
              "cookieAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/account/delete": {
        "post": {
          "summary": "Delete account and all data",
          "responses": {
            "501": {
              "description": "Not Implemented"
            }
          }
        }
      },
      "/account/create": {
        "post": {
          "summary": "Create account",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccount"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "409": {
              "description": "Conflict"
            }
          }
        }
      },
      "/account/deactivate": {
        "post": {
          "summary": "Deactivate account",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeactivateAccount"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      },
      "/account/activate": {
        "post": {
          "summary": "Activate account",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateAccount"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      }
    },
    "components": { 
      "securitySchemes": {
        "cookieAuth": {
          "type": "apiKey",
          "in": "cookie",
          "name": "cookie"
        }
      },
      "schemas": {
        "LoginRequest": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["email", "password"]
        },
        "PasswordResetRequest": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            }
          },
          "required": ["email"]
        },
        "LoginCodeRequest": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "code": {
              "type": "string"
            }
          },
          "required": ["email", "code"]
        },
        "PasswordReset": {
          "type": "object",
          "properties": {
            "new_password": {
              "type": "string"
            }
          },
          "required": ["new_password"]
        },
        "PasswordUpdate": {
          "type": "object",
          "properties": {
            "old_password": {
              "type": "string"
            },
            "new_password": {
              "type": "string"
            }
          },
          "required": ["old_password", "new_password"]
        },
        "CreateAccount": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "name": {
              "type": "string"
            }
          },
          "required": ["email", "password"]
        },
        "DeactivateAccount": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            }
          },
          "required": ["email"]
        },
        "ActivateAccount": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "code": {
              "type": "string"
            }
          },
          "required": ["email", "code"]
        }
      }
    }
  }
  